+++ sphinx/application/data/concepts/repositories/concept-repository-contact/src/main/java/chat/sphinx/concept_repository_contact/ContactRepository.kt
@@ -74,4 +74,6 @@ interface ContactRepository {
         fileName: String,
         contentLength: Long?
     ): Response<Any, ResponseError>
+
+    suspend fun toggleContactBlocked(contact: Contact): Response<Boolean, ResponseError>
 }