+++ sphinx/application/data/features/feature-coredb/src/main/java/chat/sphinx/feature_coredb/adapters/contact/ContactAdapters.kt
@@ -1,5 +1,7 @@
 package chat.sphinx.feature_coredb.adapters.contact
 
+import chat.sphinx.wrapper_common.contact.Blocked
+import chat.sphinx.wrapper_common.contact.toBlocked
 import chat.sphinx.wrapper_common.lightning.LightningNodeAlias
 import chat.sphinx.wrapper_common.lightning.LightningRouteHint
 import chat.sphinx.wrapper_contact.*
@@ -101,6 +103,27 @@ internal class PrivatePhotoAdapter: ColumnAdapter<PrivatePhoto, Long> {
     }
 
     override fun encode(value: PrivatePhoto): Long {
+        return value.value.toLong()
+    }
+}
+
+internal class BlockedAdapter private constructor(): ColumnAdapter<Blocked, Long> {
+
+    companion object {
+        @Volatile
+        private var instance: BlockedAdapter? = null
+        fun getInstance(): BlockedAdapter =
+            instance ?: synchronized(this) {
+                instance ?: BlockedAdapter()
+                    .also { instance = it }
+            }
+    }
+
+    override fun decode(databaseValue: Long): Blocked {
+        return databaseValue.toInt().toBlocked()
+    }
+
+    override fun encode(value: Blocked): Long {
         return value.value.toLong()
     }
 }