+++ sphinx/screens-detail/qr-code/qr-code/src/main/java/chat/sphinx/qr_code/ui/QRCodeFragment.kt
@@ -9,8 +9,10 @@ import androidx.lifecycle.lifecycleScope
 import by.kirich1409.viewbindingdelegate.viewBinding
 import chat.sphinx.insetter_activity.InsetterActivity
 import chat.sphinx.insetter_activity.addNavigationBarPadding
+import chat.sphinx.menu_bottom.ui.MenuBottomViewState
 import chat.sphinx.qr_code.R
 import chat.sphinx.qr_code.databinding.FragmentQrCodeBinding
+import chat.sphinx.share_qr_code.BottomMenuShareQRCode
 import dagger.hilt.android.AndroidEntryPoint
 import io.matthewnelson.android_feature_screens.ui.sideeffect.SideEffectFragment
 import io.matthewnelson.android_feature_screens.util.goneIfFalse
@@ -29,6 +31,13 @@ internal class QRCodeFragment: SideEffectFragment<
     override val binding: FragmentQrCodeBinding by viewBinding(FragmentQrCodeBinding::bind)
     override val viewModel: QRCodeViewModel by viewModels()
 
+    private val bottomMenuShareQRCode: BottomMenuShareQRCode by lazy(LazyThreadSafetyMode.NONE) {
+        BottomMenuShareQRCode(
+            onStopSupervisor,
+            viewModel
+        )
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         viewModel
@@ -40,22 +49,33 @@ internal class QRCodeFragment: SideEffectFragment<
         (requireActivity() as InsetterActivity)
             .addNavigationBarPadding(binding.layoutConstraintQrCodeFragment)
 
-        binding.includeQrCodeHeader.apply {
-            textViewDetailScreenClose.setOnClickListener {
+        binding.apply {
+            includeQrCodeHeader.textViewDetailScreenClose.setOnClickListener {
                 lifecycleScope.launch(viewModel.mainImmediate) {
                     viewModel.navigator.closeDetailScreen()
                 }
             }
 
-            textViewDetailScreenHeaderNavBack.setOnClickListener {
+            includeQrCodeHeader.textViewDetailScreenHeaderNavBack.setOnClickListener {
                 lifecycleScope.launch(viewModel.mainImmediate) {
                     viewModel.navigator.popBackStack()
                 }
             }
-        }
 
-        binding.buttonQrCodeCopy.setOnClickListener {
-            viewModel.copyCodeToClipboard()
+            buttonQrCodeCopy.setOnClickListener {
+                viewModel.copyCodeToClipboard()
+            }
+
+            buttonQrCodeShare.setOnClickListener {
+                viewModel.shareQRCodeMenuHandler.viewStateContainer.updateViewState(
+                    MenuBottomViewState.Open
+                )
+            }
+
+            bottomMenuShareQRCode.initialize(
+                includeLayoutMenuBottomShareQrCode,
+                viewLifecycleOwner
+            )
         }
     }