+++ sphinx/screens-detail/tribe-members-list/tribe-members-list/src/main/java/chat/sphinx/tribe_members_list/ui/TribeMembersListFragment.kt
@@ -4,11 +4,9 @@ import android.content.Context
 import android.os.Bundle
 import android.view.View
 import android.widget.ImageView
-import androidx.core.content.ContextCompat
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.ConcatAdapter
-import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import by.kirich1409.viewbindingdelegate.viewBinding
@@ -16,7 +14,6 @@ import chat.sphinx.concept_image_loader.ImageLoader
 import chat.sphinx.insetter_activity.InsetterActivity
 import chat.sphinx.tribe_members_list.R
 import chat.sphinx.tribe_members_list.databinding.FragmentTribeMembersListBinding
-import chat.sphinx.tribe_members_list.ui.adapter.SwipeHelper
 import chat.sphinx.tribe_members_list.ui.adapter.TribeMembersListAdapter
 import chat.sphinx.tribe_members_list.ui.adapter.TribeMembersListFooterAdapter
 import dagger.hilt.android.AndroidEntryPoint
@@ -83,16 +80,6 @@ internal class TribeMembersListFragment: SideEffectFragment<
                 }
             })
         }
-
-        context?.let {
-            val itemTouchHelper = ItemTouchHelper(object : SwipeHelper(binding.recyclerViewTribeMembersList) {
-                override fun instantiateUnderlayButton(position: Int): List<UnderlayButton> {
-                    return listOf(deleteButton(tribeMembersListAdapter, position))
-                }
-            })
-
-            itemTouchHelper.attachToRecyclerView(binding.recyclerViewTribeMembersList)
-        }
     }
 
     override suspend fun onViewStateFlowCollect(viewState: TribeMembersListViewState) {
@@ -115,22 +102,6 @@ internal class TribeMembersListFragment: SideEffectFragment<
                 )
             }
         }
-    }
-
-    private fun deleteButton(tribeMembersListAdapter: TribeMembersListAdapter, position: Int) : SwipeHelper.UnderlayButton {
-        val button = SwipeHelper.UnderlayButton(
-            requireContext(),
-            chat.sphinx.resources.R.color.badgeRed,
-            object : SwipeHelper.UnderlayButtonClickListener {
-                override fun onClick() {
-                    tribeMembersListAdapter.removeAt(position)
-                }
-            })
-        ContextCompat.getDrawable(requireContext(), R.drawable.ic_icon_delete)?.let {
-            button.addIcon(it, requireContext().resources.getDimension(R.dimen.members_list_member_holder_height))
-        }
-
-        return button
     }
 
     override suspend fun onSideEffectCollect(sideEffect: TribeMembersListSideEffect) {