+++ sphinx/screens/chats/chat-common/chat-common/src/main/java/chat/sphinx/chat_common/ui/viewstate/messageholder/LayoutState.kt
@@ -45,8 +45,10 @@ internal sealed class LayoutState private constructor() {
         val subjectName: String?,
     ): LayoutState()
 
-    data class DeletedMessage(
+    data class DeletedOrFlaggedMessage(
         val gravityStart: Boolean,
+        val deleted: Boolean,
+        val flagged: Boolean,
         val timestamp: String,
     ): LayoutState()